fix(warnings-msvc): warning C4267: 'argument': conversion from 'size_t' to 'unsigned int', possible loss of data

+++ engine/src/renderer/opengl/OpenGlBuffer.hpp
@@ -107,7 +107,7 @@ namespace nexo::renderer {
              * Usage:
              * - Use this method for dynamically updating buffer content.
              */
-            void setData(void *data, unsigned int size) override;
+            void setData(void *data, size_t size) override;
 
             [[nodiscard]] unsigned int getId() const override { return _id; };
